+++ packages/@valapi/riot-api/src/service/AccountV1.ts
@@ -31,7 +31,7 @@ export class AccountV1 extends RiotApiService {
      * @returns {Promise<AxiosResponse<AccountV1.AccountDto>>}
      */
     public byPuuid(puuid: string): Promise<AxiosResponse<AccountV1.AccountDto>> {
-        return this.axios.get(`${this.apiRegion.url.api}/riot/account/v1/accounts/by-puuid/${puuid}`);
+        return this.axios.get(`${this.apiRegion.url.continent}/riot/account/v1/accounts/by-puuid/${puuid}`);
     }
 
     /**
@@ -41,28 +41,27 @@ export class AccountV1 extends RiotApiService {
      * @returns {Promise<AxiosResponse<AccountV1.AccountDto>>}
      */
     public byRiotId(gameName: string, tagLine: string): Promise<AxiosResponse<AccountV1.AccountDto>> {
-        return this.axios.get(`${this.apiRegion.url.api}/riot/account/v1/accounts/by-riot-id/${gameName}/${tagLine}`);
+        return this.axios.get(`${this.apiRegion.url.continent}/riot/account/v1/accounts/by-riot-id/${gameName}/${tagLine}`);
     }
 
     /**
      * Get active shard for a player
      * @param {string} puuid Player UUID
-     * @param {string} game Game (default: val)
      * @returns {Promise<AxiosResponse<AccountV1.ActiveShardDto>>}
      */
-    public activeShardsByGameAndPuuid(puuid: string, game: "val" | "lor" = "val"): Promise<AxiosResponse<AccountV1.ActiveShardDto>> {
-        return this.axios.get(`${this.apiRegion.url.api}/riot/account/v1/active-shards/by-game/${game}/by-puuid/${puuid}`);
+    public activeShardsByGameAndPuuid(puuid: string): Promise<AxiosResponse<AccountV1.ActiveShardDto>> {
+        return this.axios.get(`${this.apiRegion.url.continent}/riot/account/v1/active-shards/by-game/val/by-puuid/${puuid}`);
     }
 
     /**
-     * ! Not For Public Use
+     * ! This API service is required your project to be registered by Riot Games.
      *
      * Get account by access token
      * @param {string} authorization (Header Parameters)
      * @returns {Promise<AxiosResponse<AccountV1.AccountDto>>}
      */
     public byAccessToken(authorization: string): Promise<AxiosResponse<AccountV1.AccountDto>> {
-        return this.axios.get(`${this.apiRegion.url.api}/riot/account/v1/accounts/me`, {
+        return this.axios.get(`${this.apiRegion.url.continent}/riot/account/v1/accounts/me`, {
             headers: {
                 Authorization: authorization
             }