+++ packages/@valapi/valorant-api.com/src/service/Gamemodes.ts
@@ -25,6 +25,7 @@ export namespace Gamemodes {
             state: boolean;
         }>;
         displayIcon: string;
+        listViewIconTall: string;
         assetPath: string;
     }