+++ packages/@valapi/valorant-api.com/src/service/LevelBorders.ts
@@ -1,8 +1,10 @@
 import { ValorantApiComService } from "../client/ValorantApiComService";
+import type { ValorantApiCom } from "../client/ValorantApiCom";
 
 export namespace LevelBorders {
-    export interface LevelBorders {
+    export interface LevelBorders<L extends ValorantApiCom.Language> {
         uuid: string;
+        displayName: ValorantApiComService.Languages<string, L>;
         startingLevel: number;
         levelNumberAppearance: string;
         smallPlayerCardAppearance: string;
@@ -10,12 +12,12 @@ export namespace LevelBorders {
     }
 }
 
-export class LevelBorders extends ValorantApiComService {
-    public get(): Promise<ValorantApiComService.Response<LevelBorders.LevelBorders[]>> {
+export class LevelBorders<L extends ValorantApiCom.Language = any> extends ValorantApiComService {
+    public get(): Promise<ValorantApiComService.Response<LevelBorders.LevelBorders<L>[]>> {
         return this.axios.get("/levelborders");
     }
 
-    public getByUuid(uuid: string): Promise<ValorantApiComService.Response<LevelBorders.LevelBorders>> {
+    public getByUuid(uuid: string): Promise<ValorantApiComService.Response<LevelBorders.LevelBorders<L>>> {
         return this.axios.get(`/levelborders/${uuid}`);
     }
 }