+++ packages/@valapi/web-client/src/service/Store.ts
@@ -145,7 +145,7 @@ export class Store extends WebClientService {
 
     /**
      * @param {string} subject Player UUID
-     * @param {ItemTypeId.Identify} itemTypeId Item Type
+     * @param {ItemTypeId.ID} itemTypeId Item Type
      * @returns {Promise<AxiosResponse<Store.EntitlementsWithInstance>>}
      */
     public getEntitlements(subject: string, itemTypeId: "dd3bf334-87f3-40bd-b043-682a57a8dc3a"): Promise<AxiosResponse<Store.EntitlementsWithInstance>>;