+++ packages/valorant.ts/src/utils/ValPatchNote.ts
@@ -18,7 +18,7 @@ export class ValPatchNote {
      *
      * @param {string} language Language of the patch note
      */
-    private constructor(language: Locale.Identify = Locale.Default.English_United_States) {
+    private constructor(language: Locale.ID = Locale.Default.English_United_States) {
         const _languageLowerCase = language.toLowerCase();
 
         this.url = {
@@ -30,10 +30,10 @@ export class ValPatchNote {
     /**
      *
      * @param {ValVersion.Version} version Version
-     * @param {Locale.Identify} language Language
+     * @param {Locale.ID} language Language
      * @returns {string} Url
      */
-    public static getUrl(version?: ValVersion.Version, language?: Locale.Identify): string {
+    public static getUrl(version?: ValVersion.Version, language?: Locale.ID): string {
         const patchNoteInstance = new ValPatchNote(language);
 
         if (!version) {
@@ -45,11 +45,11 @@ export class ValPatchNote {
 
     /**
      *
-     * @param {Locale.Identify} language Language
+     * @param {Locale.ID} language Language
      * @param {AxiosRequestConfig} axiosConfig Config
      * @returns {Promise<AxiosResponse<any>>} Patch Note Content
      */
-    public static async getContent(language?: Locale.Identify, axiosConfig?: AxiosRequestConfig): Promise<AxiosResponse<any>> {
+    public static async getContent(language?: Locale.ID, axiosConfig?: AxiosRequestConfig): Promise<AxiosResponse<any>> {
         const patchNoteInstance = new ValPatchNote(language);
 
         return axios.get(`${patchNoteInstance.url.pageData}/news/tags/patch-notes/page-data.json`, axiosConfig);