+++ packages/@valapi/riot-api/CHANGELOG.md
@@ -1,3 +1,35 @@
+# 4.0.0
+
+**Change**
+
+-   ~~`RiotApiRegion.riotRegion`~~ **-->** `RiotApiRegion.continent`
+-   ~~`RiotApiRegion.url.api`~~ **-->** `RiotApiRegion.url.continent`
+-   ~~`RiotApiRegion.url.server`~~ **-->** `RiotApiRegion.url.region`
+
+### Typescript
+
+**Add**
+
+-   `MatchV1.MatchInfoDto`
+-   `MatchV1.AbilityCastsDto`
+-   `MatchV1.PlayerStatsDto`
+-   `MatchV1.PlayerDto`
+-   `MatchV1.CoachDto`
+-   `MatchV1.TeamDto`
+-   `MatchV1.LocationDto`
+-   `MatchV1.PlayerLocationsDto`
+-   `MatchV1.FinishingDamageDto`
+-   `MatchV1.KillDto`
+-   `MatchV1.DamageDto`
+-   `MatchV1.EconomyDto`
+-   `MatchV1.AbilityDto`
+-   `MatchV1.PlayerRoundStatsDto`
+-   `MatchV1.RoundResultDto`
+-   `MatchV1.MatchDto`
+-   `MatchV1.MatchlistEntryDto`
+-   `MatchV1.MatchlistDto`
+-   `MatchV1.RecentMatchesDto`
+
 # 3.2.0
 
 **Add**